requirements_telemetry.txt
# OpenTelemetry Requirements for IMPALA RL

# Core OpenTelemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# OTLP Exporters (for SigNoz)
opentelemetry-exporter-otlp-proto-grpc>=1.20.0

# Instrumentation
opentelemetry-instrumentation>=0.41b0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# Optional: For enhanced trace propagation
opentelemetry-propagator-b3>=1.20.0
